g: check the "remote" name in the branch description
When the head commit of the branch has the message starting with
"[name/name ...]", the current code misinterprets the first and second
"name"s to be the remote name and the remote branch name.

To avoid this situation, we introduce a check for the remote name in
this patch.  We interpret the first name to be the remote name only
when it matches a remote name already registered in the repository.